fix to work with floating point numbers

+++ main.rs
@@ -11,7 +11,7 @@ fn main() {
     io::stdin()
         .read_line(&mut op)
         .expect("Failed to read line");
-    let op: i32 = op.trim().parse().expect("Select number of your operation.");
+    let op: f32 = op.trim().parse().expect("Select number of your operation.");
 
     if op == 1 {
         println!("Enter your first number to be added:");
@@ -19,14 +19,14 @@ fn main() {
         io::stdin()
             .read_line(&mut x)
             .expect("Failed to read line");
-        let x: i32 = x.trim().parse().expect("Enter your first number to be added:");
+        let x: f32 = x.trim().parse().expect("Enter your first number to be added:");
     
         println!("Enter your second number:");
         let mut y = String::new();
         io::stdin()
             .read_line(&mut y)
             .expect("Failed to read line");
-        let y: i32 = y.trim().parse().expect("Enter your second number:");
+        let y: f32 = y.trim().parse().expect("Enter your second number:");
         add(x,y); 
     } else if op == 2 {
         println!("Enter your first number to be subtracted: (Format num1-num2)");
@@ -34,14 +34,14 @@ fn main() {
         io::stdin()
             .read_line(&mut x)
             .expect("Failed to read line");
-        let x: i32 = x.trim().parse().expect("Enter your first number to be subtracted: (Format num1-num2)");
+        let x: f32 = x.trim().parse().expect("Enter your first number to be subtracted: (Format num1-num2)");
     
         println!("Enter your second number:");
         let mut y = String::new();
         io::stdin()
             .read_line(&mut y)
             .expect("Failed to read line");
-        let y: i32 = y.trim().parse().expect("Enter your second number:");
+        let y: f32 = y.trim().parse().expect("Enter your second number:");
         sub(x,y); 
     } else if op == 3 {
         println!("Enter your first number to multiply:");
@@ -49,14 +49,14 @@ fn main() {
         io::stdin()
             .read_line(&mut x)
             .expect("Failed to read line");
-        let x: i32 = x.trim().parse().expect("Enter your first number to multiply:");
+        let x: f32 = x.trim().parse().expect("Enter your first number to multiply:");
     
         println!("Enter your second number:");
         let mut y = String::new();
         io::stdin()
             .read_line(&mut y)
             .expect("Failed to read line");
-        let y: i32 = y.trim().parse().expect("Enter your second number:");
+        let y: f32 = y.trim().parse().expect("Enter your second number:");
         mult(x,y); 
     } else if op == 4 {
         println!("Enter your first number to divide: (Format: num1/num2)");
@@ -64,14 +64,14 @@ fn main() {
         io::stdin()
             .read_line(&mut x)
             .expect("Failed to read line");
-        let x: i32 = x.trim().parse().expect("Enter your first number to divide: (Format: num1/num2)");
+        let x: f32 = x.trim().parse().expect("Enter your first number to divide: (Format: num1/num2)");
     
         println!("Enter your second number:");
         let mut y = String::new();
         io::stdin()
             .read_line(&mut y)
             .expect("Failed to read line");
-        let y: i32 = y.trim().parse().expect("Enter your second number:");
+        let y: f32 = y.trim().parse().expect("Enter your second number:");
         div(x,y); 
     } else {
         println!("Invalid input.");
@@ -79,22 +79,22 @@ fn main() {
 
 }
 
-fn sub(ado:i32, adt:i32) {
+fn sub(ado:f32, adt:f32) {
     let ada = ado-adt;
     println!("{}", ada)
 }
 
-fn div(ado:i32, adt:i32) {
+fn div(ado:f32, adt:f32) {
     let ada = ado/adt;
     println!("{}", ada)
 }
 
-fn mult(ado:i32, adt:i32) {
+fn mult(ado:f32, adt:f32) {
     let ada = ado*adt;
     println!("{}", ada)
 }
 
-fn add(ado:i32, adt:i32) {
+fn add(ado:f32, adt:f32) {
     let ada = ado+adt;
     println!("{}", ada)
 }
